Document npm install workaround
We depend on better-sqlite3@7.6.2, which has a build issue which can be worked around as described.

better-sqlite is not actually needed, it's just a transitive depedency: secp256k1_hash_to_curve_circom > circom-helper > better-sqlite3@7.6.2

The best solution is to edit secp256k1_hash_to_curve_circom so that we can import the circuits and typescript helpers without also importing the testing infrastructure, which better-sqlite3 is needed for. The workaround should be redundant once the dependencies are fixed.