`rust-arkworks` `hash_to_curve` reference `fn` upgrade
It was transferred to appropriate location, relevant dependepcies
updated, code was improved and reworked.
Find below some rationale on the most discussionable choice
of the rework.

## Rationale
`from_random_bytes` might be preferred way that will be found
in the actual hash_to_curve functions, but it doesn't let us be generic
in this reference example. `[&u8]` interface is broad and makes too few
assumptions to be sure it's able to preserve properties of our function.
Adding that behind broad interface only shallow trait definition, it's
hard to be sure in its behavior in different contexts. These design
choices could be justified to discourage its usage out of context
of particular (curve) implementation.

That said, devs using the trait should be seeking standardized
primitives, and delve into all the details down the chain
of calls when it's unavoidable.

I feel like input randomness could be damaged in subtle ways using
`from_random_bytes` without thorough understanding of particular
implementation. Future updates can damage it too
in non-obvious way as crate(s) isn't quite stable yet.

So as far as it's just an example *let's go with naive but transparent
approach.* Complexity of production implemetation can grow to very
significant level as seen at
<https://docs.rs/ark-ec/0.4.2/ark_ec/hashing/curve_maps/swu/trait.SWUConfig.html>.